IDEMPIERE-5567 Support of UUID as Key (FHCA-4195)
- Manage translation for UUID based tables
- Enable 2Pack Import UUID based tables with translation
- Create ad_package_imp_detail.Record_UU
- Create test table TestUU_Trl

+++ org.adempiere.pipo.handlers/src/org/adempiere/pipo2/handler/IndexColumnElementHandler.java
@@ -54,8 +54,8 @@ public class IndexColumnElementHandler extends AbstractElementHandler {
 			MIndexColumn mIndexColumn = findPO(ctx, element);
 			if (mIndexColumn == null) {
 				int parentId = 0;
-				if (getParentId(element, MTableIndex.Table_Name) > 0) {
-					parentId = getParentId(element, MTableIndex.Table_Name);
+				if ((Integer)getParentId(element, MTableIndex.Table_Name) > 0) {
+					parentId = (Integer)getParentId(element, MTableIndex.Table_Name);
 				} else {
 					Element pfElement = element.properties.get(MIndexColumn.COLUMNNAME_AD_TableIndex_ID);
 					parentId = ReferenceUtils.resolveReferenceAsInt(ctx.ctx, pfElement, getTrxName(ctx));
